+++ include/MemoryDiagnostics.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <Arduino.h>
+
+namespace MemoryDiagnostics {
+void log(const char *label);
+}